refactor auth page
